Add stuff

+++ src/kvraft/common.go
@@ -13,7 +13,7 @@ type PutAppendArgs struct {
 	Key   string
 	Value string
 	Op    string // "Put" or "Append"
-	ClerkID int64
+	ClerkID int
 	RequestID int
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
@@ -26,7 +26,7 @@ type PutAppendReply struct {
 
 type GetArgs struct {
 	Key string
-	ClerkID int64 
+	ClerkID int
 	RequestID int
 	// You'll have to add definitions here.
 }
